refactored components folder + Products page
