+++ src/components/commons/StampCard.tsx
