+++ src/pages/about_us.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import StyledImg from '@/components/StyledImg'
+import StyledImg from '@/components/commons/StyledImg'
 import { Typography, Grid } from '@mui/material'
 import storytext from '../data/aboutUsStory';
 import CenteredWrapper from '@/components/layouts/CenteredWrapper';
